// src/interfaces/RectangleOptions.ts
/**
 * @interface RectangleOptions - Rectangle options
 * @property {string} [fillColor] - Fill color
 * @property {number} [weight] - Stroke weight
 * @property {string} [strokeColor] - Stroke color
 * @public
 */
export interface RectangleOptions {
  fillColor?: string;
  weight?: number;
  strokeColor?: string;
}